fix frontend bugs and revise home and reviewer profile page

+++ resume_review/models.py
@@ -49,7 +49,7 @@ class Reviewer(models.Model):
         ('delivery_3', 'Three weeks'),
         ('delivery_4', 'Four weeks or more'),
     ]
-    rate = models.IntegerField(default=0)
+    rate = models.FloatField(default=0)
     delivery_time = models.CharField(
         max_length=255, choices=DELIVERY_TIME_CHOICES, default='delivery_1')
